+++ src/App.tsx
@@ -36,6 +36,8 @@ import HqCountryDashboard from './pages/hq/CountryDashboard'
 import HqApplications from './pages/hq/Applications'
 import HqRequestsLeader from './pages/hq/RequestsLeader'
 import HqRequestsPartnerByLeader from './pages/hq/RequestsPartnerByLeader'
+import HqRequestsPartnerDirect from './pages/hq/RequestsPartnerDirect'
+import HqRequestsMerchantDirect from './pages/hq/RequestsMerchantDirect'
 import HqLeaders from './pages/hq/Leaders'
 import HqPartners from './pages/hq/Partners'
 import HqMerchants from './pages/hq/Merchants'
@@ -62,6 +64,8 @@ const HQ_PAGES: Record<string, JSX.Element> = {
   applications: <HqApplications />,
   'requests/leader': <HqRequestsLeader />,
   'requests/partner-by-leader': <HqRequestsPartnerByLeader />,
+  'requests/partner-direct': <HqRequestsPartnerDirect />,
+  'requests/merchant-direct': <HqRequestsMerchantDirect />,
   leaders: <HqLeaders />,
   'leaders/sales': <HqLeaderSales />,
   partners: <HqPartners />,